Updated migrations for travis fix

+++ db/migrate/20201107203204_create_review_bids.rb
@@ -1,15 +1,17 @@
 class CreateReviewBids < ActiveRecord::Migration[4.2]
+  # will break
   def change
     create_table :review_bids do |t|
       t.integer  'priority',       limit: 4
-      t.integer  'signuptopic_id', limit: 4
-      t.integer  'participant_id', limit: 4
+      t.bigint  'signuptopic_id', limit: 4
+      t.bigint  'participant_id', limit: 4
       t.datetime 'created_at'
       t.datetime 'updated_at'
-      t.integer  'user_id',        limit: 4
-      t.integer  'assignment_id',  limit: 4
+      t.bigint  'user_id',        limit: 4
+      t.bigint  'assignment_id',  limit: 4
       t.timestamps null: false
     end
+    # t.references :assignment, index: true, foreign_key: true
 
     add_index 'review_bids', ['assignment_id'], name: 'fk_rails_549e23ae08', using: :btree
     add_index 'review_bids', ['participant_id'], name: 'fk_rails_ab93feeb35', using: :btree
